email_automation/requirements.txt
Django>=4.2

google-auth
google-auth-oauthlib
google-api-python-client

python-dotenv
